[Ares][Pushdown] Separate time filtering from row filtering and needed refactoring
Summary:
It turns out that Ares wants to have rowFilters not contain the
time clause. The time clause should only be in the timeFilters section
of the query. The row filters should contain only the "remaining
filters".

So we need to use the remainingPredicate for the filter AND the
remaining column domains. So I had to use the code that converts a
pushdown expression tree out of the tuple domain (of the other non time
columns).

We also handle disjuncts in the time filter domain: the remainder holes
are put back into the row filters.

(Alternative approach would have been to teach the DomainExtractor to
only focus on one column, but that was a bit more scary change).

So the refactoring is to pull up the tuple domain -> expression tree
into a utility SPI class out of the Pinot module. This refactoring
required some unit test changes.

Also, we needed to know the Type of the symbols in the tuple domain
pushed into the filter, so we expose that into the filter node as well.

Test Plan: Unit tests

Reviewers: venkik, vivekb, bhasudha

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3034981