[Ares] Time series caching at a split level
Summary:
Cache the results of each ares split if deemed cacheable.
Create the splits in such a way that they are UTC day aligned. Aligned
splits help get more hits for the "bulk" of the time region. This
approach is a simplification of that used in Pythia.

The caching is on demand with a configurable max memory map (per connector for
now). New regions (ie those less than configurable time) from the current time
are not cached because we expect that data to change.

Also added metrics for the cache. We could use a similar approach for
pinot segments even.

Depends on https://code.uberinternal.com/D3034981

Test Plan: Tested manually using overlapping intervals

Reviewers: burak, venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3035075