[Neutrino][Ares] Simplify the multithreaded aresdb splitting within the ares page source
Summary:
get rid of the extra single thread executor. It messes up the
jstack with these unnamed threads and is kind of not needed since it was
reading a future and populating a linkedblockingqueue. Instead just
transform the future and get it directly.

Also add some session log entries, clearly associating the ares calls
with the split

Test Plan:
Ran queries, checked that the session log entries and the
output data are correct.

Reviewers: ujwala, venkik, vivekb

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3067415