[Neutrino] Expose the total time that a query is blocked on scan
Summary:
For neutrino, it is important to track the "non scan" time. Assume
that the only time a query spends is doing scan, what is the remainder of the
time ? This time represents everything from presto-useful work time (like doing
valid work like aggregation etc) to scheduling slack.

I will create an alert on this metric for neutrino. Also log this to heatpipe.

Test Plan: Ran on neutrino and saw this metric appear on grafana.

Reviewers: lu.niu, venkik, bhasudha, curt

Reviewed By: curt

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3083295